requirements.txt
flask==2.0.1
google-cloud-firestore==2.1.1
scapy==2.4.5